Minor updates for lua 5.3
Simple changes to all the LUA_VERSION_NUM == 502 tests throughout the
code to make them >= 502, to allow for lua 5.3.  In addition, lua 5.3
changed the signature for lua_dump() to add a strip parameter.  In
those cases, call the new signature with strip = 0 to retain the old
behavior.

This is necessary to properly build on centos 8, Fedora 31 and above
with lua 5.3.  Based on a patch in the Fedora tree by Breno Brand
Fernandes brandfbb@gmail.com which replaced the problematic
calls with one for the new version of lua.  This patch maintains
backwards compatibility with older versions of lua, should that
be necessary.