Define __P() macro in libvbr
Definitions for this macro have been added throughout the codebase in
commits 91e7407d, 705948fd, 227fa252, 842c1733, and b730bdc0, but one
was still missing from libvbr. glibc contains a definition for legacy
reasons, but other libcs might not. Particularly, the musl libc does not
contain it, leading to build errors when enabling support for VBR.

Add a definition for __P() to vbr.h to fix this.